requirements.txt
# FastAPI app (Py3.11-compatible pins w/ NumPy 1.x)
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0

# Pydantic v2 (FastAPI 0.110 expects this)
pydantic==2.6.4
pydantic-core==2.16.3

# Scientific stack (lock to NumPy 1.x to satisfy SciPy & scikit-learn)
numpy==1.26.4
scipy==1.11.4
pandas==2.2.1
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.3.2

# Extras
orjson==3.9.15
prometheus-fastapi-instrumentator==7.0.0
redis==5.0.1
requests==2.31.0

azure-storage-blob